Update lv_obj_style.c
Fix type error under idf 6.0 in log warning

+++ src/core/lv_obj_style.c
@@ -336,7 +336,7 @@ void lv_obj_style_set_disabled(lv_obj_t * obj, const lv_style_t * style, lv_styl
             return;
         }
     }
-    LV_LOG_WARN("%p style was not found on %p widget with %6x selector", (void *)style, (void *)obj, selector);
+    LV_LOG_WARN("%p style was not found on %p widget with %6x selector", (void *)style, (void *)obj, (int) selector);
 }
 
 bool lv_obj_style_get_disabled(lv_obj_t * obj, const lv_style_t * style, lv_style_selector_t selector)
@@ -348,7 +348,7 @@ bool lv_obj_style_get_disabled(lv_obj_t * obj, const lv_style_t * style, lv_styl
         }
     }
 
-    LV_LOG_WARN("%p style was not found on %p widget with %6x selector", (void *)style, (void *)obj, selector);
+    LV_LOG_WARN("%p style was not found on %p widget with %6x selector", (void *)style, (void *)obj, (int) selector);
     return false;
 }
